Completed work on new connections format. Testing script included.
New system uses a flag named "default" to know which is the default connections along with the code to auto name clusters/nodes when the name flag is not given.
connection_suite.py is the script containing all the testing information. Checks all the push and remove connections I could come up with.